added chrome web-store trial period setup post
This blog post will be helpful for those who are looking to set up a free trial period for their Chrome extension hosted in the Chrome web store. I added this link because several bits of information is missing from the official Chrome extension documentation.

I'm not sure if this PR  is qualified for this repo. So feel free to close this PR.